+++ finality-aleph/src/config.rs
@@ -1,0 +1,20 @@
+use sp_core::traits::BareCryptoStorePtr;
+use std::time::Duration;
+
+/// Configuration for the Aleph protocol service.
+pub struct Config {
+    /// The duration of a message to be sent across the network.
+    pub gossip_duration: Duration,
+    /// If the node is running as an authority.
+    pub is_authority: bool,
+    /// The name of this particular node.
+    pub name: Option<String>,
+    /// The keystore which stores the keys.
+    pub keystore: Option<BareCryptoStorePtr>,
+}
+
+impl Config {
+    pub(crate) fn name(&self) -> &str {
+        self.name.as_deref().unwrap_or("<unknown>")
+    }
+}
